+++ src/main/java/io/github/squid233/decoration/block/ModBlocks.java
@@ -18,11 +18,11 @@ import java.util.Locale;
 public enum ModBlocks implements ModBlockConvertible {
     PLATFORM_1(new PlatformBlock(AbstractBlock.Settings.copy(Blocks.SMOOTH_STONE))),
     PLATFORM_2(new Block(AbstractBlock.Settings.copy(Blocks.SMOOTH_STONE))),
-    PANTOGRAPH(new PantographBlock(AbstractBlock.Settings.copy(Blocks.BLACK_CONCRETE).nonOpaque())),
+    PANTOGRAPH(new PantographBlock(AbstractBlock.Settings.copy(Blocks.BLACK_CONCRETE))),
     WIRE_POLE(new WirePoleBlock(AbstractBlock.Settings.copy(Blocks.LIGHT_GRAY_CONCRETE))),
     CATENARY_POLE(new CatenaryPoleBlock(AbstractBlock.Settings.copy(Blocks.LIGHT_GRAY_CONCRETE))),
     CATENARY_BI_POLE(new CatenaryBiPoleBlock(AbstractBlock.Settings.copy(Blocks.LIGHT_GRAY_CONCRETE))),
-    CATENARY_CROSS_POLE(new CatenaryCrossPoleBlock(AbstractBlock.Settings.copy(Blocks.LIGHT_GRAY_CONCRETE).nonOpaque())),
+    CATENARY_CROSS_POLE(new CatenaryCrossPoleBlock(AbstractBlock.Settings.copy(Blocks.LIGHT_GRAY_CONCRETE))),
     CATENARY_POLE_EXTRA(new CatenaryPoleExtraBlock(AbstractBlock.Settings.copy(Blocks.LIGHT_GRAY_CONCRETE))),
     CATENARY_PART(new CatenaryPartBlock(AbstractBlock.Settings.copy(Blocks.LIGHT_GRAY_CONCRETE))),
     TRAFFIC_LIGHT_3(new TrafficLightBlock.Light3(AbstractBlock.Settings.copy(Blocks.BLACK_CONCRETE).luminance(value -> 3))),