+++ src/main/java/io/github/squid233/decoration/block/entity/ModBlockEntityTypes.java
@@ -13,8 +13,8 @@ import net.minecraft.registry.Registry;
 public final class ModBlockEntityTypes {
     public static final BlockEntityType<CatenaryPartBlockEntity> CATENARY_PART_BLOCK_ENTITY_TYPE =
         BlockEntityType.Builder.create(CatenaryPartBlockEntity::new, ModBlocks.CATENARY_PART.toBlock()).build(null);
-    public static final BlockEntityType<TrafficLight3BlockEntity> TRAFFIC_LIGHT_3_BLOCK_ENTITY_TYPE =
-        BlockEntityType.Builder.create(TrafficLight3BlockEntity::new, ModBlocks.TRAFFIC_LIGHT_3.toBlock()).build(null);
+    public static final BlockEntityType<TrafficLightBlockEntity.Light3> TRAFFIC_LIGHT_3_BLOCK_ENTITY_TYPE =
+        BlockEntityType.Builder.create(TrafficLightBlockEntity.Light3::new, ModBlocks.TRAFFIC_LIGHT_3.toBlock()).build(null);
 
     private static void register(String name, BlockEntityType<?> type) {
         Registry.register(Registries.BLOCK_ENTITY_TYPE, Decoration.id(name), type);