+++ src/main/java/io/github/squid233/decoration/client/render/block/entity/TrafficLight3BlockEntityRenderer.java
@@ -2,8 +2,8 @@ package io.github.squid233.decoration.client.render.block.entity;
 
 import io.github.squid233.decoration.block.ModBlocks;
 import io.github.squid233.decoration.block.TrafficLightBlock;
-import io.github.squid233.decoration.block.entity.TrafficLight3BlockEntity;
-import io.github.squid233.decoration.client.DecorationClient;
+import io.github.squid233.decoration.block.entity.TrafficLightBlockEntity;
+import io.github.squid233.decoration.block.entity.TrafficLightStep;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.block.BlockState;
@@ -19,39 +19,22 @@ import net.minecraft.world.World;
 import org.joml.Matrix4f;
 import org.joml.Quaternionf;
 
+import java.util.List;
+
 /**
  * @author squid233
  * @since 0.1.0
  */
 @Environment(EnvType.CLIENT)
-public class TrafficLight3BlockEntityRenderer implements BlockEntityRenderer<TrafficLight3BlockEntity> {
+public class TrafficLight3BlockEntityRenderer implements BlockEntityRenderer<TrafficLightBlockEntity.Light3> {
     private static final float epsilon = 1.0e-4f;
     private static final Quaternionf QUAT = new Quaternionf();
 
     public TrafficLight3BlockEntityRenderer(BlockEntityRendererFactory.Context context) {
     }
 
-    private enum Color {
-        NONE(0, 0, 0, 0),
-        RED(255, 0, 0, 11f / 16),
-        YELLOW(255, 216, 0, 6f / 16),
-        GREEN(0, 255, 33, 1f / 16);
-
-        private final int red;
-        private final int green;
-        private final int blue;
-        private final float yOffset;
-
-        Color(int red, int green, int blue, float yOffset) {
-            this.red = red;
-            this.green = green;
-            this.blue = blue;
-            this.yOffset = yOffset;
-        }
-    }
-
     @Override
-    public void render(TrafficLight3BlockEntity entity, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, int overlay) {
+    public void render(TrafficLightBlockEntity.Light3 entity, float tickDelta, MatrixStack matrices, VertexConsumerProvider vertexConsumers, int light, int overlay) {
         final World world = entity.getWorld();
         if (world == null) {
             return;
@@ -65,53 +48,62 @@ public class TrafficLight3BlockEntityRenderer implements BlockEntityRenderer<Tra
 
         final Direction direction = blockState.get(TrafficLightBlock.FACING);
 
-        final long time = world.getTime() % (DecorationClient.trafficLightRedTicks + DecorationClient.trafficLightYellowTicks * 2L + DecorationClient.trafficLightGreenTicks);
-
-        matrices.push();
-        matrices.translate(0.5f, 0.0f, 0.5f);
-        matrices.multiply(QUAT.rotationY((float) Math.toRadians(-direction.asRotation())));
-        matrices.translate(-0.125f, 0.0f, -0.125f + epsilon);
-        final MatrixStack.Entry peek = matrices.peek();
-
-        final VertexConsumer buffer = vertexConsumers.getBuffer(RenderLayer.getDebugQuads());
-
-        final Color color = determineColor(time, direction);
-        if (color != Color.NONE) {
-            renderQuad(peek, buffer,
-                color.yOffset,
-                color.yOffset + (4f / 16),
-                color.red, color.green, color.blue);
+        final var steps = entity.steps();
+        long totalTicks = 0;
+        for (TrafficLightStep step : steps) {
+            totalTicks += step.ticks();
         }
 
-        matrices.pop();
+        if (totalTicks > 0) {
+            final long time = world.getTime() % totalTicks;
+
+            matrices.push();
+            matrices.translate(0.5f, 0.0f, 0.5f);
+            matrices.multiply(QUAT.rotationY((float) Math.toRadians(-direction.asRotation())));
+            matrices.translate(-0.125f, 0.0f, -0.125f + epsilon);
+            final MatrixStack.Entry peek = matrices.peek();
+
+            final VertexConsumer buffer = vertexConsumers.getBuffer(RenderLayer.getDebugQuads());
+
+            final var step = determineStep(time, steps);
+            final var color = step.color();
+            if (color.isColor()) {
+                final Integer colorValue = color.getColorValue();
+                if (colorValue != null) {
+                    final float yOffset = computeYOffset(step.index());
+                    renderQuad(peek, buffer,
+                        yOffset,
+                        yOffset + (4f / 16),
+                        0xff000000 | colorValue);
+                }
+            }
+
+            matrices.pop();
+        }
     }
 
-    private static Color determineColor(long time, Direction direction) {
-        final int redTicks = DecorationClient.trafficLightRedTicks;
-        final int yellowTicks = DecorationClient.trafficLightYellowTicks;
-        final int greenTicks = DecorationClient.trafficLightGreenTicks;
-        if (time < redTicks) {
-            if (direction.getAxis() == Direction.Axis.X) {
-                return Color.RED;
+    private static float computeYOffset(int index) {
+        return (1 + index * 5) / 16f;
+    }
+
+    private static TrafficLightStep determineStep(long time, List<TrafficLightStep> steps) {
+        long ticks = 0;
+        for (TrafficLightStep step : steps) {
+            final int ticks1 = step.ticks();
+            if (ticks1 <= 0) {
+                continue;
             }
-            return Color.GREEN;
-        }
-        if (time < redTicks + yellowTicks) {
-            if (time % 20 < 10) {
-                return Color.YELLOW;
+            ticks += ticks1;
+            if (time < ticks) {
+                final int flashing = step.flashing();
+                if (flashing == 0 ||
+                    time % ((double) ticks1 / flashing) < ticks1 * 0.5 / flashing) {
+                    return step;
+                }
+                return TrafficLightStep.EMPTY;
             }
-            return Color.NONE;
         }
-        if (time < redTicks + yellowTicks + greenTicks) {
-            if (direction.getAxis() == Direction.Axis.X) {
-                return Color.GREEN;
-            }
-            return Color.RED;
-        }
-        if (time % 20 < 10) {
-            return Color.YELLOW;
-        }
-        return Color.NONE;
+        return TrafficLightStep.EMPTY;
     }
 
     private static void renderQuad(
@@ -119,12 +111,12 @@ public class TrafficLight3BlockEntityRenderer implements BlockEntityRenderer<Tra
         VertexConsumer buffer,
         float y0,
         float y1,
-        int red, int green, int blue
+        int colorValue
     ) {
         final Matrix4f positionMatrix = entry.getPositionMatrix();
-        buffer.vertex(positionMatrix, 0.0f, y1, 0.0f).color(red, green, blue, 255).next();
-        buffer.vertex(positionMatrix, 0.0f, y0, 0.0f).color(red, green, blue, 255).next();
-        buffer.vertex(positionMatrix, 0.25f, y0, 0.0f).color(red, green, blue, 255).next();
-        buffer.vertex(positionMatrix, 0.25f, y1, 0.0f).color(red, green, blue, 255).next();
+        buffer.vertex(positionMatrix, 0.0f, y1, 0.0f).color(colorValue).next();
+        buffer.vertex(positionMatrix, 0.0f, y0, 0.0f).color(colorValue).next();
+        buffer.vertex(positionMatrix, 0.25f, y0, 0.0f).color(colorValue).next();
+        buffer.vertex(positionMatrix, 0.25f, y1, 0.0f).color(colorValue).next();
     }
 }