+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.2.0
+
+- Changed
+    - Traffic light
+        - Now use a config screen
+- Removed
+    - Game rules: `trafficLight*Ticks`
+
 # 0.1.0
 
 - Added